pyproject.toml
[tool.poetry]
name = "dreamai"
version = "3.0"
description = ""
authors = ["Hamza Farhan <thehamza96@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<=3.12"
notebook = "^7.2.1"
ipython = "^8.25.0"
ipywidgets = "^8.1.3"
pytest = "^8.2.2"
ruff = "^0.4.9"
termcolor = "^2.4.0"
python-dotenv = "^1.0.1"
instructor = "^1.3.3"
openai = "^1.34.0"
anthropic = "^0.28.1"
google-generativeai = "^0.7.2"
google-auth-oauthlib = "^1.2.0"
fastcore = "^1.5.45"
tiktoken = "^0.7.0"
demoji = "^1.1.0"
langgraph = "^0.0.69"
requests = "^2.32.3"
numpy = "1.26.4"
pandas = "^2.2.2"
sentence-transformers = "^3.0.1"
lancedb = "^0.12.0"
burr = {extras = ["start"], version = "^0.22.1"}
graphviz = "^0.20.3"
click = "^8.1.7"
loguru = "^0.7.2"
uvicorn = "^0.30.1"
fastapi = "^0.111.0"
aiofiles = "^23.2.1"
pre-commit = "^3.7.1"
firecrawl-py = "^0.0.20"
tantivy = "^0.22.0"
exa-py = "^1.0.12"
langchain = "^0.2.11"
langchain-community = {extras = ["pypdf"], version = "^0.2.10"}
pypdf = "^4.3.1"
gliner = "^0.2.8"
sentencepiece = "^0.2.0"
instructorembedding = "^1.0.1"
pymupdf4llm = "^0.0.10"
openparse = "^0.5.7"
duckduckgo-search = "^6.2.6"
tavily-python = "^0.3.8"
pymongo = "^4.8.0"
pydantic-settings = "^2.4.0"
python-fasthtml = "^0.4.5"
html2text = "^2024.2.26"
lxml = "^5.3.0"

[tool.ruff]
line-length = 95

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
